code cleanup and effecting suggested changes
