+++ src/common/groupedOrdersTable.scss
@@ -1,7 +1,7 @@
 @use '@carbon/layout';
 @use '@carbon/colors';
 @use '@openmrs/esm-styleguide/src/vars' as *;
-@use '@carbon/styles/scss/spacing';
+@use '@carbon/layout';
 @use '@carbon/styles/scss/type';
 
 .tableContainer {
