+++ src/common/listOrderDetails.scss
@@ -50,6 +50,6 @@
   
   .actionBtns {
     float: right;
-    margin-left: 1rem;
-    margin-top: 2rem;
+    margin-left: layout.$spacing-05;
+    margin-top: layout.$spacing-07;
   }
